Refactor old warehousetest
